+++ java-r5rcore/src/org/ipea/r5r/AccessibilityEstimator.java
@@ -1,32 +1,19 @@
 package org.ipea.r5r;
 
 import com.conveyal.r5.OneOriginResult;
-import com.conveyal.r5.analyst.FreeFormPointSet;
 import com.conveyal.r5.analyst.PointSet;
 import com.conveyal.r5.analyst.TravelTimeComputer;
 import com.conveyal.r5.analyst.cluster.RegionalTask;
 import com.conveyal.r5.analyst.decay.*;
-import com.conveyal.r5.api.util.LegMode;
-import com.conveyal.r5.profile.StreetMode;
 import com.conveyal.r5.transit.TransportNetwork;
-import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
 
 public class AccessibilityEstimator extends R5MultiDestinationProcess {
 
-    private static final org.slf4j.Logger LOG = LoggerFactory.getLogger(AccessibilityEstimator.class);
-
-    private FreeFormPointSet destinationPoints;
     private DecayFunction decayFunction;
 
     public void setDecayFunction(String decayFunction, double decayValue) {
@@ -51,63 +38,19 @@ public class AccessibilityEstimator extends R5MultiDestinationProcess {
 
     public AccessibilityEstimator(ForkJoinPool threadPool, TransportNetwork transportNetwork, RoutingProperties routingProperties) {
         super(threadPool, transportNetwork, routingProperties);
-        destinationPoints = null;
     }
 
     @Override
-    public List<LinkedHashMap<String, ArrayList<Object>>> run() throws ExecutionException, InterruptedException {
-        buildDestinationPointSet();
-        return super.run();
-    }
-
-    private void buildDestinationPointSet() {
-        ByteArrayOutputStream dataStream = new ByteArrayOutputStream();
-        DataOutputStream pointStream = new DataOutputStream(dataStream);
-
-        try {
-            pointStream.writeInt(toIds.length);
-            for (String toId : toIds) {
-                pointStream.writeUTF(toId);
-            }
-            for (double toLat : toLats) {
-                pointStream.writeDouble(toLat);
-            }
-            for (double toLon : toLons) {
-                pointStream.writeDouble(toLon);
-            }
-            for (int opportunity : opportunities) {
-                pointStream.writeDouble(opportunity);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        ByteArrayInputStream pointsInput = new ByteArrayInputStream(dataStream.toByteArray());
-
-        try {
-            destinationPoints = new FreeFormPointSet(pointsInput);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        if (!this.directModes.isEmpty()) {
-            for (LegMode mode : this.directModes) {
-                transportNetwork.linkageCache.getLinkage(destinationPoints, transportNetwork.streetLayer, StreetMode.valueOf(mode.toString()));
-            }
-        }
-    }
-
-    @Override
-    protected LinkedHashMap<String, ArrayList<Object>> runProcess(int index) throws ParseException {
+    protected RDataFrame runProcess(int index) throws ParseException {
         RegionalTask request = buildRequest(index);
 
         TravelTimeComputer computer = new TravelTimeComputer(request, transportNetwork);
         OneOriginResult travelTimeResults = computer.computeTravelTimes();
-        RDataFrame travelTimesTable = buildDataFrameStructure(fromIds[index]);
+        RDataFrame travelTimesTable = buildDataFrameStructure(fromIds[index], 10);
         populateDataFrame(travelTimeResults, travelTimesTable);
 
         if (travelTimesTable.nRow() > 0) {
-            return travelTimesTable.getDataFrame();
+            return travelTimesTable;
         } else {
             return null;
         }
@@ -130,9 +73,10 @@ public class AccessibilityEstimator extends R5MultiDestinationProcess {
         }
     }
 
-    private RDataFrame buildDataFrameStructure(String fromId) {
+    @Override
+    protected RDataFrame buildDataFrameStructure(String fromId, int nRows) {
         // Build return table
-        RDataFrame travelTimesTable = new RDataFrame();
+        RDataFrame travelTimesTable = new RDataFrame(nRows);
         travelTimesTable.addStringColumn("from_id", fromId);
         travelTimesTable.addIntegerColumn("percentile", 0);
         travelTimesTable.addIntegerColumn("cutoff", 0);