+++ r-package/R/accessibility.R
@@ -343,8 +343,10 @@ accessibility <- function(r5r_core,
   # process results ---------------------------------------------------------
 
   # convert travel_times from java object to data.table
+  cat("Preparing final output...")
   accessibility <- jdx::convertToR(accessibility)
-  accessibility <- data.table::rbindlist(accessibility)
+  data.table::setDT(accessibility)
+  cat(" DONE!\n")
 
   return(accessibility)
 }