+++ r-package/R/travel_time_matrix.R
@@ -191,7 +191,7 @@ travel_time_matrix <- function(r5r_core,
   departure <- posix_to_string(departure_datetime)
 
   # max trip duration
-  checkmate::assert_numeric(max_trip_duration)
+  checkmate::assert_numeric(max_trip_duration, lower=1)
   max_trip_duration <- as.integer(max_trip_duration)
 
   # max_walking_distance, max_bike_distance, and max_street_time
@@ -205,10 +205,12 @@ travel_time_matrix <- function(r5r_core,
   # origins and destinations
   origins      <- assert_points_input(origins, "origins")
   destinations <- assert_points_input(destinations, "destinations")
-
-
+  
+  checkmate::assert_subset("id", names(origins))
+  checkmate::assert_subset("id", names(destinations))
+  
   # time window
-  checkmate::assert_numeric(time_window)
+  checkmate::assert_numeric(time_window, lower=1)
   time_window <- as.integer(time_window)
   draws <- time_window *5
   draws <- as.integer(draws)
@@ -266,8 +268,10 @@ travel_time_matrix <- function(r5r_core,
   # process results ---------------------------------------------------------
 
   # convert travel_times from java object to data.table
+  cat("Preparing final output...")
+
   travel_times <- jdx::convertToR(travel_times)
-  travel_times <- data.table::rbindlist(travel_times)
+  data.table::setDT(travel_times)
 
   # convert eventual list columns to integer
   for(j1 in seq_along(travel_times)) {
@@ -281,5 +285,6 @@ travel_time_matrix <- function(r5r_core,
     data.table::set(travel_times, i=which(travel_times[[j]]>max_trip_duration), j=j, value=NA_integer_)
     }
 
-return(travel_times)
+  cat(" DONE!\n")
+  return(travel_times)
 }