+++ r-package/NEWS.md
@@ -1,7 +1,21 @@
 # log history of r5r package development
 
 -------------------------------------------------------
-# r5r v0.5.1 (dev)
+# r5r v0.6.0 (dev)
+
+**Major changes**
+
+- Updated R5 to version 6.4. Closes [#182](https://github.com/ipeaGIT/r5r/issues/182).
+
+- Significant performance improvements in all functions, due to a faster method
+for consolidating outputs. Closes [#180](https://github.com/ipeaGIT/r5r/issues/180)
+
+- New function `transit_network_to_sf()`, to extract the public transport network 
+from R5 as simple features. Issue [#179](https://github.com/ipeaGIT/r5r/issues/179)
+
+**Minor changes**
+
+- New `setup_r5()` parameter, `overwrite`, that forces the building of a new `network.dat`, even if one already exists.
 
 
 -------------------------------------------------------